ath79: Forward port support for CR3000 based on ar71xx
The PowerCloud Systems CR3000 was a cloud-managed CPE for
a now defunct NaaS offering.  It was previously supported
under the ar71xx branch and this forward ports that support
with some notable differences:

1) Since reverting to stock firmware is now irrelevant there is
is only a single openwrt image generated which uses the entire
flash rather than preserving PowerCloud-specific partitions
that are unneeded to openwrt -- those partitions will be erased
and used by the openwrt image.

2) Rather than use a non-standard probe order for the ethernet
devices, this image uses a set of 'ip link set ethX name ethY'
commands very early in preinit (before the network is used at all),
in order to have the the switch and Wan use the same ethernet names
as in previous images.

3) /etc/config/wireless will need to be regenerated as the path
to the wireless device has changed due to differences in ath79
DT for ar93x compared to ar71xx images.

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>